+++ examples/default/main.tf
@@ -1,12 +1,12 @@
 provider "aws" {
   region  = "eu-west-1"
-  version = "1.39"
+  version = "2.1.0"
 }
 
 provider "aws" {
   alias   = "us-east-1"
   region  = "us-east-1"
-  version = "1.39"
+  version = "2.1.0"
 }
 
 locals {