+++ output.tf
@@ -12,3 +12,12 @@ output "cloudfront_bucket_arn" {
   description = "ARN of the bucket for serving content."
   value       = "${aws_s3_bucket.web.arn}"
 }
+
+output "cloudfront_distribution" {
+  description = "Cloudfront distribution (id, arn)"
+
+  value = {
+    id  = "${aws_cloudfront_distribution.web.id}"
+    arn = "${aws_cloudfront_distribution.web.arn}"
+  }
+}